+++ test/unordered_map_test.cpp
@@ -2,7 +2,7 @@
  * Copyright (C) 2019  Hariharan Devarajan, Keith Bateman
  *
  * This file is part of HCL
- * 
+ *
  * HCL is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
  * published by the Free Software Foundation, either version 3 of the
@@ -101,7 +101,7 @@ int main (int argc,char* argv[])
     if (debug) {
         printf("%s/%d: %d\n", processor_name, my_rank, getpid());
     }
-    
+
     if(debug && my_rank==0){
         printf("%d ready for attach\n", comm_size);
         fflush(stdout);
@@ -109,7 +109,7 @@ int main (int argc,char* argv[])
     }
     MPI_Barrier(MPI_COMM_WORLD);
     bool is_server=(my_rank+1) % ranks_per_server == 0;
-    int my_server=my_rank / ranks_per_server;
+    size_t my_server=my_rank / ranks_per_server;
     int num_servers=comm_size/ranks_per_server;
 
     // The following is used to switch to 40g network on Ares.
@@ -122,7 +122,7 @@ int main (int argc,char* argv[])
 
     size_t size_of_elem = sizeof(int);
 
-    printf("rank %d, is_server %d, my_server %d, num_servers %d\n",my_rank,is_server,my_server,num_servers);
+    printf("rank %d, is_server %d, my_server %zu, num_servers %d\n",my_rank,is_server,my_server,num_servers);
 
     const int array_size=TEST_REQUEST_SIZE;
 
@@ -132,7 +132,7 @@ int main (int argc,char* argv[])
 
     std::array<int,array_size> my_vals=std::array<int,array_size>();
 
-    
+
     HCL_CONF->IS_SERVER = is_server;
     HCL_CONF->MY_SERVER = my_server;
     HCL_CONF->NUM_SERVERS = num_servers;
